improve the status and validation

+++ database/migrations/2024_10_22_083236_reviews.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->string('review');
-            $table->float('stars',2);
+            $table->integer('stars');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('business_id')->constrained('business')->onDelete('cascade');
             $table->timestamps();
